+++ db/yas_db_database.cpp
@@ -207,8 +207,8 @@ class db::database::impl : public base::impl {
                 data = "";
             }
             sqlite3_bind_blob(stmt, column_idx, data, static_cast<int>(blob.size()), SQLITE_STATIC);
-        } else if (type == typeid(db::int64)) {
-            sqlite3_bind_int64(stmt, column_idx, value.value<db::int64>());
+        } else if (type == typeid(db::integer)) {
+            sqlite3_bind_int64(stmt, column_idx, value.value<db::integer>());
         } else if (type == typeid(db::float64)) {
             sqlite3_bind_double(stmt, column_idx, value.value<db::float64>());
         } else if (type == typeid(db::string)) {
